Added workloadAssignedQueues logic handling it. (#8433)
* Added workloadAssignedQueues logic handling it.

* Misspell correction.

* Unchecked responses error fix (cache and cache_test)

* Post merge cleanup.

* Update pkg/cache/scheduler/cache.go


* SImplified assigsnment treacking after assumption removal and rebase.

* Logic error change - fix.

* FOrget workload change revert.

* Naming fix.

* Unified and updated the "add or update workload (withour lock)" logic.

* Fized error in update logic. Restructured code for readability.

* Code readability improvements.

* Added test case for unassigning workload quota.

* Tests: AddOrUpdateWorkload method name fix in test names.

* Added missing error check.

* Refactored addOrUpdateWorkloadWithoutLock. Updated when method says an update has occured.

* Major code cleanup.

* Update pkg/cache/scheduler/cache.go


---------

+++ pkg/cache/scheduler/cache_test.go
@@ -1221,11 +1221,12 @@ func TestCacheWorkloadOperations(t *testing.T) {
 	}
 
 	steps := []struct {
-		name           string
-		operation      func(log logr.Logger, cache *Cache) error
-		wantResults    map[kueue.ClusterQueueReference]result
-		wantError      string
-		wantLocalQueue queue.LocalQueueReference
+		name               string
+		operation          func(log logr.Logger, cache *Cache) error
+		wantResults        map[kueue.ClusterQueueReference]result
+		wantError          string
+		wantLocalQueue     queue.LocalQueueReference
+		wantQueuesAssigned map[workload.Reference]kueue.ClusterQueueReference
 	}{
 		{
 			name: "add",
@@ -1308,16 +1309,101 @@ func TestCacheWorkloadOperations(t *testing.T) {
 			},
 		},
 		{
+			name: "UpdateWorkload; quota assigned -> quota unassigned",
+			operation: func(log logr.Logger, cache *Cache) error {
+				w := utiltestingapi.MakeWorkload("b", "").Obj()
+				if err := cache.UpdateWorkload(log, w); err != nil {
+					return errors.Join(errors.New("failed to update workload"), err)
+				}
+				return nil
+			},
+			wantResults: map[kueue.ClusterQueueReference]result{
+				"one": {
+					Workloads: sets.New[workload.Reference]("/a"),
+					UsedResources: resources.FlavorResourceQuantities{
+						{Flavor: "on-demand", Resource: corev1.ResourceCPU}: 10,
+						{Flavor: "spot", Resource: corev1.ResourceCPU}:      15,
+					},
+				},
+				"two": {
+					Workloads: sets.New[workload.Reference]("/c"),
+				},
+			},
+		},
+		{
+			name: "UpdateWorkload; quota not assigned -> quota still not assigned",
+			operation: func(log logr.Logger, cache *Cache) error {
+				w := utiltestingapi.MakeWorkload("d", "").Obj()
+				if err := cache.UpdateWorkload(log, w); err != nil {
+					return errors.Join(errors.New("failed to update workload"), err)
+				}
+				return nil
+			},
+			wantResults: map[kueue.ClusterQueueReference]result{
+				"one": {
+					Workloads: sets.New[workload.Reference]("/a", "/b"),
+					UsedResources: resources.FlavorResourceQuantities{
+						{Flavor: "on-demand", Resource: corev1.ResourceCPU}: 10,
+						{Flavor: "spot", Resource: corev1.ResourceCPU}:      15,
+					},
+				},
+				"two": {
+					Workloads: sets.New[workload.Reference]("/c"),
+				},
+			},
+		},
+		{
+			name: "AddOrUpdateWorkload; quota assigned -> quota unassigned",
+			operation: func(log logr.Logger, cache *Cache) error {
+				w := utiltestingapi.MakeWorkload("b", "").Obj()
+				if cache.AddOrUpdateWorkload(log, w) {
+					return errors.New("declared workload update performed when only a deletion should have been performed")
+				}
+				return nil
+			},
+			wantResults: map[kueue.ClusterQueueReference]result{
+				"one": {
+					Workloads: sets.New[workload.Reference]("/a"),
+					UsedResources: resources.FlavorResourceQuantities{
+						{Flavor: "on-demand", Resource: corev1.ResourceCPU}: 10,
+						{Flavor: "spot", Resource: corev1.ResourceCPU}:      15,
+					},
+				},
+				"two": {
+					Workloads: sets.New[workload.Reference]("/c"),
+				},
+			},
+		},
+		{
+			name: "AddOrUpdateWorkload; quota not assigned -> quota still not assigned",
+			operation: func(log logr.Logger, cache *Cache) error {
+				w := utiltestingapi.MakeWorkload("d", "").Obj()
+				if cache.AddOrUpdateWorkload(log, w) {
+					return errors.New("declared workload update performed when no action should have been taken")
+				}
+				return nil
+			},
+			wantResults: map[kueue.ClusterQueueReference]result{
+				"one": {
+					Workloads: sets.New[workload.Reference]("/a", "/b"),
+					UsedResources: resources.FlavorResourceQuantities{
+						{Flavor: "on-demand", Resource: corev1.ResourceCPU}: 10,
+						{Flavor: "spot", Resource: corev1.ResourceCPU}:      15,
+					},
+				},
+				"two": {
+					Workloads: sets.New[workload.Reference]("/c"),
+				},
+			},
+		},
+		{
 			name: "update cluster queue for a workload",
 			operation: func(log logr.Logger, cache *Cache) error {
-				old := utiltestingapi.MakeWorkload("a", "").ReserveQuotaAt(&kueue.Admission{
-					ClusterQueue: "one",
-				}, now).Obj()
 				latest := utiltestingapi.MakeWorkload("a", "").PodSets(podSets...).ReserveQuotaAt(&kueue.Admission{
 					ClusterQueue:      "two",
 					PodSetAssignments: psAssignments,
 				}, now).Obj()
-				return cache.UpdateWorkload(log, old, latest)
+				return cache.UpdateWorkload(log, latest)
 			},
 			wantResults: map[kueue.ClusterQueueReference]result{
 				"one": {
@@ -1337,42 +1423,14 @@ func TestCacheWorkloadOperations(t *testing.T) {
 			},
 		},
 		{
-			name: "update error old clusterQueue doesn't exist",
+			name: "update error new ClusterQueue not found",
 			operation: func(log logr.Logger, cache *Cache) error {
-				old := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
-					ClusterQueue: "three",
-				}, now).Obj()
-				latest := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
-					ClusterQueue: "one",
-				}, now).Obj()
-				return cache.UpdateWorkload(log, old, latest)
-			},
-			wantError: "old ClusterQueue doesn't exist",
-			wantResults: map[kueue.ClusterQueueReference]result{
-				"one": {
-					Workloads: sets.New[workload.Reference]("/a", "/b"),
-					UsedResources: resources.FlavorResourceQuantities{
-						{Flavor: "on-demand", Resource: corev1.ResourceCPU}: 10,
-						{Flavor: "spot", Resource: corev1.ResourceCPU}:      15,
-					},
-				},
-				"two": {
-					Workloads: sets.New[workload.Reference]("/c"),
-				},
-			},
-		},
-		{
-			name: "update error new clusterQueue doesn't exist",
-			operation: func(log logr.Logger, cache *Cache) error {
-				old := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
-					ClusterQueue: "one",
-				}, now).Obj()
 				latest := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
 					ClusterQueue: "three",
 				}, now).Obj()
-				return cache.UpdateWorkload(log, old, latest)
+				return cache.UpdateWorkload(log, latest)
 			},
-			wantError: "new ClusterQueue doesn't exist",
+			wantError: "cluster queue not found",
 			wantResults: map[kueue.ClusterQueueReference]result{
 				"one": {
 					Workloads: sets.New[workload.Reference]("/a", "/b"),
@@ -1389,13 +1447,10 @@ func TestCacheWorkloadOperations(t *testing.T) {
 		{
 			name: "update workload which doesn't exist.",
 			operation: func(log logr.Logger, cache *Cache) error {
-				old := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
-					ClusterQueue: "one",
-				}, now).Obj()
 				latest := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
 					ClusterQueue: "two",
 				}, now).Obj()
-				return cache.UpdateWorkload(log, old, latest)
+				return cache.UpdateWorkload(log, latest)
 			},
 			wantResults: map[kueue.ClusterQueueReference]result{
 				"one": {
@@ -1451,12 +1506,11 @@ func TestCacheWorkloadOperations(t *testing.T) {
 			},
 		},
 		{
-			name: "attempt deleting non-existing workload with cancelled admission",
+			name: "delete non-existing workload with cancelled admission",
 			operation: func(log logr.Logger, cache *Cache) error {
 				w := utiltestingapi.MakeWorkload("d", "").Obj()
 				return cache.DeleteWorkload(log, w)
 			},
-			wantError: "cluster queue not found",
 			wantResults: map[kueue.ClusterQueueReference]result{
 				"one": {
 					Workloads: sets.New[workload.Reference]("/a", "/b"),
@@ -1473,9 +1527,25 @@ func TestCacheWorkloadOperations(t *testing.T) {
 		{
 			name: "delete error clusterQueue doesn't exist",
 			operation: func(log logr.Logger, cache *Cache) error {
-				w := utiltestingapi.MakeWorkload("a", "").ReserveQuotaAt(&kueue.Admission{
+				cq := utiltestingapi.MakeClusterQueue("three").
+					ResourceGroup(
+						*utiltestingapi.MakeFlavorQuotas("on-demand").Resource("cpu").Obj(),
+						*utiltestingapi.MakeFlavorQuotas("spot").Resource("cpu").Obj(),
+					).
+					NamespaceSelector(nil).
+					Obj()
+				w := utiltestingapi.MakeWorkload("d", "").ReserveQuotaAt(&kueue.Admission{
 					ClusterQueue: "three",
 				}, now).Obj()
+
+				if err := cache.AddClusterQueue(logr.NewContext(t.Context(), log), cq); err != nil {
+					return err
+				}
+				if updated, _ := cache.addOrUpdateWorkloadWithoutLock(log, w); !updated {
+					return errors.New("failed to add test workload")
+				}
+				cache.DeleteClusterQueue(cq)
+
 				return cache.DeleteWorkload(log, w)
 			},
 			wantError: "cluster queue not found",
@@ -1490,6 +1560,12 @@ func TestCacheWorkloadOperations(t *testing.T) {
 				"two": {
 					Workloads: sets.New[workload.Reference]("/c"),
 				},
+			},
+			wantQueuesAssigned: map[workload.Reference]kueue.ClusterQueueReference{
+				"/a": "one",
+				"/b": "one",
+				"/c": "two",
+				"/d": "three",
 			},
 		},
 		{
@@ -1525,6 +1601,16 @@ func TestCacheWorkloadOperations(t *testing.T) {
 				}
 			}
 
+			wantQueuesAssigned := step.wantQueuesAssigned
+			if wantQueuesAssigned == nil {
+				wantQueuesAssigned = map[workload.Reference]kueue.ClusterQueueReference{}
+				for cq, res := range step.wantResults {
+					for wlKey := range res.Workloads {
+						wantQueuesAssigned[wlKey] = cq
+					}
+				}
+			}
+
 			gotError := step.operation(log, cache)
 			if diff := cmp.Diff(step.wantError, messageOrEmpty(gotError)); diff != "" {
 				t.Errorf("Unexpected error (-want,+got):\n%s", diff)
@@ -1538,6 +1624,9 @@ func TestCacheWorkloadOperations(t *testing.T) {
 			}
 			if diff := cmp.Diff(step.wantResults, gotResult, cmpopts.EquateEmpty()); diff != "" {
 				t.Errorf("Unexpected clusterQueues (-want,+got):\n%s", diff)
+			}
+			if diff := cmp.Diff(wantQueuesAssigned, cache.workloadAssignedQueues, cmpopts.EquateEmpty()); diff != "" {
+				t.Errorf("Unexpected clusterQueues assignments for workloads (-want,+got):\n%s", diff)
 			}
 		})
 	}
@@ -2843,7 +2932,7 @@ func TestCachePodsReadyForAllAdmittedWorkloads(t *testing.T) {
 					Type:   kueue.WorkloadPodsReady,
 					Status: metav1.ConditionTrue,
 				})
-				return cache.UpdateWorkload(log, wl, newWl)
+				return cache.UpdateWorkload(log, newWl)
 			},
 			wantReady: true,
 		},
@@ -2866,7 +2955,7 @@ func TestCachePodsReadyForAllAdmittedWorkloads(t *testing.T) {
 					Type:   kueue.WorkloadPodsReady,
 					Status: metav1.ConditionFalse,
 				})
-				return cache.UpdateWorkload(log, wl, newWl)
+				return cache.UpdateWorkload(log, newWl)
 			},
 			wantReady: false,
 		},
@@ -2893,7 +2982,7 @@ func TestCachePodsReadyForAllAdmittedWorkloads(t *testing.T) {
 					Type:   kueue.WorkloadPodsReady,
 					Status: metav1.ConditionTrue,
 				})
-				return cache.UpdateWorkload(log, wl2, newWl2)
+				return cache.UpdateWorkload(log, newWl2)
 			},
 			wantReady: true,
 		},
